+++ src/arcane_arcade_spelleffects/TeleportEffect.java
@@ -47,19 +47,19 @@ public class TeleportEffect extends SpellEffect
 	// IMPLEMENTED METHODS	----------------------------------------------
 
 	@Override
-	public void onBallCollision(Ball ball, int x, int y)
+	public void onBallCollision(Ball ball, double x, double y)
 	{
 		// Doesn't collide with balls
 	}
 
 	@Override
-	public void onSpellCollision(SpellEffect spell, int x, int y)
+	public void onSpellCollision(SpellEffect spell, double x, double y)
 	{
 		// Doesn't collide with spells
 	}
 
 	@Override
-	public void onWizardCollision(Wizard wizard, int x, int y)
+	public void onWizardCollision(Wizard wizard, double x, double y)
 	{
 		// Doesn't collide with wizards
 	}
